requirements.txt
SQLAlchemy
Pillow
python-dotenv
